follow nextjs route gen strat
