+++ apps/sim/app/api/workspaces/[id]/pdf/preview/route.ts
@@ -1,67 +1,14 @@
-import { createLogger } from '@sim/logger'
-import { type NextRequest, NextResponse } from 'next/server'
-import { getSession } from '@/lib/auth'
-import { toError } from '@/lib/core/utils/helpers'
-import { MAX_DOCUMENT_PREVIEW_CODE_BYTES } from '@/lib/execution/constants'
-import { runSandboxTask } from '@/lib/execution/sandbox/run-task'
-import { verifyWorkspaceMembership } from '@/app/api/workflows/utils'
+import { createDocumentPreviewRoute } from '@/app/api/workspaces/[id]/_preview/create-preview-route'
 
 export const dynamic = 'force-dynamic'
 export const runtime = 'nodejs'
-
-const logger = createLogger('PdfPreviewAPI')
 
 /**
  * POST /api/workspaces/[id]/pdf/preview
  * Compile PDF-Lib source code and return the binary PDF for streaming preview.
  */
-export async function POST(req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
-  const { id: workspaceId } = await params
-
-  try {
-    const session = await getSession()
-    if (!session?.user?.id) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
-    }
-
-    const membership = await verifyWorkspaceMembership(session.user.id, workspaceId)
-    if (!membership) {
-      return NextResponse.json({ error: 'Insufficient permissions' }, { status: 403 })
-    }
-
-    let body: unknown
-    try {
-      body = await req.json()
-    } catch {
-      return NextResponse.json({ error: 'Invalid or missing JSON body' }, { status: 400 })
-    }
-    const { code } = body as { code?: string }
-
-    if (typeof code !== 'string' || code.trim().length === 0) {
-      return NextResponse.json({ error: 'code is required' }, { status: 400 })
-    }
-
-    if (Buffer.byteLength(code, 'utf-8') > MAX_DOCUMENT_PREVIEW_CODE_BYTES) {
-      return NextResponse.json({ error: 'code exceeds maximum size' }, { status: 413 })
-    }
-
-    const buffer = await runSandboxTask(
-      'pdf-generate',
-      { code, workspaceId },
-      { ownerKey: `user:${session.user.id}`, signal: req.signal }
-    )
-
-    return new NextResponse(new Uint8Array(buffer), {
-      status: 200,
-      headers: {
-        'Content-Type': 'application/pdf',
-        'Content-Length': String(buffer.length),
-        'Cache-Control': 'private, no-store',
-      },
-    })
-  } catch (err) {
-    const message = toError(err).message
-    logger.error('PDF preview generation failed', { error: message, workspaceId })
-    return NextResponse.json({ error: message }, { status: 500 })
-  }
-}
+export const POST = createDocumentPreviewRoute({
+  taskId: 'pdf-generate',
+  contentType: 'application/pdf',
+  label: 'PDF',
+})
